Refactor: Objectify
